Fixed broken tests.

+++ setup_schedule_test.go
@@ -47,9 +47,9 @@ func TestSetupSchedule(t *testing.T) {
 	assert.Contains(t, recorder.Body.String(), "Failed to publish matches")
 	assert.Nil(t, err)
 	assert.Equal(t, 64, len(matches))
-	assert.Equal(t, 1388595600, matches[0].Time.Unix())
-	assert.Equal(t, 1388685360, matches[7].Time.Unix())
-	assert.Equal(t, 1388782800, matches[24].Time.Unix())
+	assert.Equal(t, int64(1388595600), matches[0].Time.Unix())
+	assert.Equal(t, int64(1388685360), matches[7].Time.Unix())
+	assert.Equal(t, int64(1388782800), matches[24].Time.Unix())
 }
 
 func TestSetupScheduleErrors(t *testing.T) {
